nros::prelude

Enum ParameterValue

pub enum ParameterValue {
    NotSet,
    Bool(bool),
    Integer(i64),
    Double(f64),
    String(String<nros_params::::types::ParameterValue::{constant#0}>),
    ByteArray(Vec<u8, nros_params::::types::ParameterValue::{constant#1}>),
    BoolArray(Vec<bool, nros_params::::types::ParameterValue::{constant#2}>),
    IntegerArray(Vec<i64, nros_params::::types::ParameterValue::{constant#3}>),
    DoubleArray(Vec<f64, nros_params::::types::ParameterValue::{constant#4}>),
    StringArray(Vec<String<nros_params::::types::ParameterValue::{constant#5}>, nros_params::::types::ParameterValue::{constant#6}>),
}

Parameter value container

Holds a parameter value of any supported type. Note: This enum is intentionally large to support all parameter types without heap allocation in embedded systems.

Variants

NotSet

Value not set

Bool(bool)

Boolean value

Integer(i64)

64-bit signed integer value

Double(f64)

64-bit floating point value

String(String<nros_params::::types::ParameterValue::{constant#0}>)

String value

ByteArray(Vec<u8, nros_params::::types::ParameterValue::{constant#1}>)

Byte array value

BoolArray(Vec<bool, nros_params::::types::ParameterValue::{constant#2}>)

Boolean array value

IntegerArray(Vec<i64, nros_params::::types::ParameterValue::{constant#3}>)

Integer array value

DoubleArray(Vec<f64, nros_params::::types::ParameterValue::{constant#4}>)

Double array value

StringArray(Vec<String<nros_params::::types::ParameterValue::{constant#5}>, nros_params::::types::ParameterValue::{constant#6}>)

String array value (array of fixed-size strings)

Implementations

impl ParameterValue

pub fn param_type(&self) -> ParameterType

Get the parameter type for this value

pub fn is_set(&self) -> bool

Check if the value is set

pub fn as_bool(&self) -> Option<bool>

Try to get the value as a bool

pub fn as_integer(&self) -> Option<i64>

Try to get the value as an integer

pub fn as_double(&self) -> Option<f64>

Try to get the value as a double

pub fn as_string(&self) -> Option<&str>

Try to get the value as a string slice

pub fn as_byte_array(&self) -> Option<&[u8]>

Try to get the value as a byte array slice

pub fn as_bool_array(&self) -> Option<&[bool]>

Try to get the value as a bool array slice

pub fn as_integer_array(&self) -> Option<&[i64]>

Try to get the value as an integer array slice

pub fn as_double_array(&self) -> Option<&[f64]>

Try to get the value as a double array slice

pub fn from_bool(value: bool) -> ParameterValue

Create a bool value

pub fn from_integer(value: i64) -> ParameterValue

Create an integer value

pub fn from_double(value: f64) -> ParameterValue

Create a double value

pub fn from_string(value: &str) -> Option<ParameterValue>

Create a string value from a str slice