nros

Trait ParameterVariant

pub trait ParameterVariant: Clone {
    // Required methods
    fn to_parameter_value(&self) -> ParameterValue;
    fn from_parameter_value(value: &ParameterValue) -> Option<Self>;
    fn parameter_type() -> ParameterType;
}

Trait for types that can be used as typed parameters

This trait provides conversions between Rust types and ParameterValue, enabling type-safe parameter access.

Required Methods

fn to_parameter_value(&self) -> ParameterValue

Convert this type to a ParameterValue

fn from_parameter_value(value: &ParameterValue) -> Option<Self>

Try to extract this type from a ParameterValue

fn parameter_type() -> ParameterType

Get the expected parameter type

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.